Add tox bandit support for python-magnumclient
Add bandit support for python-magnumclient executed via tox.  The
idea is to gate on bandit once we work out the kinks.  In the short
term, we will have an experiemtal job that can be run manually.  After
that it will be added as a non-voting check job.  Finally it will be
set as a voting check/gate.

Change-Id: I14d037cf611ea6efa9f8271e6e120cd11b764dc4
Partially-Implements: blueprint gate-bandit