Fix translation setup
The infra scripts really expect $REPODIR/locale to work, fix
setup.cfg and rename the locale file so that infra scripts work.

This fixes change Id9f5b08835a4780a8448c989a878c03b70a97803 which
already said:"The proper place is for all python projects
$PROJECT/locale/$PROJECT.pot - see setup.cfg." - but then during review
we missed that instead of PROJECT=python-magnumclient, it used
magnumclient - and the infra scripts do not cope with that since
PROJECT really has to be the name of the repository.

Change-Id: Idc5efe52a1ff1b0cb8599e1258e411522547f0e4