Rename image_id to image when create a container
Currently, magnum used "image_id" to reflect the image in container,
but actually, we can support use both image name and image id to create
a container. The name "image_id" may confuse user that we only support
"image_id" when create a container.

This patch rename the "image_id"  to "image" when create a container.

Change-Id: I6de28ec714f5a03fd5a44c1c95e0f68fcf2856b4
Closes-Bug: #1466278