Add certificates operations
This commit adds two operations to the magnum client:

magnum ca-show --bay bay_uuid
magnum ca-sign --bay bay_uuid --csr /path/to/csr.pem

ca-show retrieves the CA certificate for the provided bay.

ca-sign, sends the provided CSR to Magnum, and prints the signed
certificate returned by Magnum. The certificate is signed with the
CA for the given Bay.

Change-Id: I784a1b3dc77e72dfb9e7f8d25cbbc37a0b5ffce0
Partial-Implements: blueprint magnum-as-a-ca