Magnum show cmds display dictionaries with unicode u chars
When a Magnum resource has an attribute that is implemented as a dictionary,
and that dictionary contains either unicode keys or unicode values, then
the Magnum client show commands display the dictionary keys/values
including a preceding 'u' character. For example, the output of the
'magnum baymodel-show <baymodel>' command includes this line:

| labels | {u'flannel_network_subnetlen': u'26',
            u'flannel_network_cidr': u'10.101.0.0/16',
            u'flannel_use_vxlan': u'true'} |

This patchset adds a recursive conversion of a dictionary's keys and
values from unicode to strings before the contents of the dictionary
are printed by the Magnum client.

Change-Id: I0fdc1f1d2dfc693652307ac66622c24f03899f8d
Closes-Bug: #1498071