Add unversioned client constructor
It's a very common pattern in OpenStack to have a base level client
file that has a client constructor that takes a version as the first
parameter. This allows developers to write code that just imports the
client, rather than relying on python import paths for version
selection, which in turn allows for future version autodetection with
things like microversions.

Closes-Bug: #1514732

Change-Id: Ibf8df9c8dbdfe27d9a387304e05afe95d3e33f49