Remove node object from magnumclient
The node object represents either a bare metal or virtual machine
node that is provisioned with an OS to run the containers, or
alternatively, run kubernetes. Magnum use Heat to deploy the nodes,
so it is unnecessary to maintain node object in Magnum. Heat can do
the work for us. The code about node object is useless now, so let's
remove it from Magnum.

Change-Id: Icb7bd21032b592c94033d93e22f460d8c2270fbe
Closes-Bug: #1541255
Depends-On: If8761b06a364127683099afb4dc51ea551be6f89