Remove bandit.yaml in favor of defaults
This patch removes the bandit.yaml and just uses whatever Bandit
defaults with in the scan. This cuts down on the maintenance of
the burdensome bandit.yaml file.

This patch also changes the severity level of the scan to just
medium and high issues.

Here are the results without the -ll to exclude low severity:
http://paste.openstack.org/show/489898/

Change-Id: Ia063db516175f518cfa1c68902b8b74734f294c0