Prints  '-' instead of 'None' when data is None
Client prints 'None' when data is None , and client also prints
'None' when data is a string 'None'. But string 'None' is  different
to None. To clear the confusion, Client should print '-' instead of
'None' when data is None. Nova Client and Cinder Client has cleared
the confusion.

Change-Id: I5b7013eb7ccfb99658702066191feed93f82fb8a
Related-Bug: #1600553
Closed-Bug: #1422244