Add shell command bay-config
Generate the required configuration for the corresponding native client
of a given bay. Output the appropriate export or setenv commands with
the native client's environment variables, ready for clients to source.

In the case of kubernetes also generate a config file which is pointed
by the KUBECONFIG env var. In the case of docker the config file does
not support this and all goes into env vars.

Change-Id: I44700b97ba3efc4c818112c95f80adf723048c5a
Implements: blueprint magnum-coe-client-config