Updates certificate CLI to use cluster
Updates commands to the CA CLI to use cluster in addition to to bay.
Adds tests for cluster commands.
Currently checks for required arguments in logic instead of through
cliutils.py. Changes to cliutils.py seem to be necessary to allow
for "either-or" required arguments.
Depends on change #354436.

Change-Id: Id8c72151661514c3a80ae9fd64369b00a5b78cc9
Implements: blueprint rename-bay-to-cluster