Add missing options for HTTPClient if auth_token given.
The insecure client options and timeout were bypassed
if auth token was given, as in case of Magnum UI.
This is causing SSL error even insecure is set to true
by Magnum UI and magnum client is unable to skip SSL check
becuase insecure option was not passed to HTTPClient.
This was caused due to this change :
https://github.com/openstack/python-magnumclient/commit/a374b9f99cabd78f93603860425517d45b6290fa#diff-c8d8ebbd1f15e85c914906c6357e0a5f
As a consequence, any HTTPS based connections to magnum endpoint
fail. With this patch, the missing options to HTTPClient are
correctly passed to and SSL checks are skipped safely if insecure
is set to True.

Change-Id: I885301d6eb69182f109cdfbfbc771bbbf13beb2a
Closes-Bug: #1634305