Use assert_called_*/assert_not_called to verify mock calls
Currently assertTrue and assertFalse are used to check
wheter a mock method is called. It is not checked whether
mock method is called with correct set of arguments.
With this patch following modules uses assert_called_* and
assert_not_called instead of assertTrue and assertFalse.

- magnumclient/tests/v1/test_certificates_shell.py
- magnumclient/tests/v1/test_mservices_shell.py

Also, renamed "file_mock" to "mock_file" for consistent
mock naming in the modules.

Change-Id: I706b362509f6a59be13df8d9c132e4f4b68e0060
Partially-Implements: blueprint magnumclient-ut-coverage