Add CT tags argument and field to client
We noticed that from the user perspective it is hard
to know when a cluster_template provided by the cloud
admin is mature enough for a production release.
This field will allow the administrator to add an
annotation to the cluster template like
{deprecated, recommended, testing} giving further
usefull information to the end user about the
template's life cycle

This patch allows the magnumclient to pass the
--tags <text> when creating a cluster template
and also shows the observations available when listing
existent cluster templates.

story: 2004215
task: 40161

Depends-On: I5d1c4221f089bc5cd12b25f620aa01771a029df9
Change-Id: Ibbdc9a9cd7ae4733eb6d38a7db11a68866f0d446
Signed-off-by: Diogo Guerra <dy090.guerra@gmail.com>